requirements.txt
pydantic-ai-slim[openai]
gradio>=6.7.0
python-dotenv>=1.0.1
ddgs>=9.0.0
httpx>=0.27.0
